pause-select/main_test.go
package main

import (
	"testing"

	"github.com/stretchr/testify/assert"
	"go.uber.org/atomic"
)

func TestMain(t *testing.T) {
	tests := []struct {
		msg string
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.msg, func(t *testing.T) {
			main()
		})
	}
}

func TestRun(t *testing.T) {
	tests := []struct {
		msg     string
		wantErr string
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.msg, func(t *testing.T) {
			err := run()
			if tt.wantErr != "" {
				assert.EqualError(t, err, tt.wantErr)
			}
		})
	}
}

func TestLockdownSignal(t *testing.T) {
	type args struct {
		lockdown  *atomic.Bool
		lockdownC chan struct{}
	}
	tests := []struct {
		msg  string
		args args
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.msg, func(t *testing.T) {
			lockdownSignal(tt.args.lockdown, tt.args.lockdownC)
		})
	}
}

func TestCheckLockdownCondition(t *testing.T) {
	type args struct {
	}
	tests := []struct {
		msg           string
		isFilePresent bool
		wantBoolValue bool
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.msg, func(t *testing.T) {
			b := atomic.NewBool(false)

			checkLockdownCondition(b)
			assert.Equal(t, b.Load(), tt.wantBoolValue)
		})
	}
}
